chore: Refactor into integration tests
