+++ src/errors.rs
@@ -1,7 +1,8 @@
+use std::fmt;
 use thiserror::Error;
 use crate::location::Location;
 
-#[derive(Error, Debug)]
+#[derive(Error)]
 pub enum MomoaError {
     #[error("Unexpected character {c:?} found.")]
     UnexpectedCharacter {
@@ -11,5 +12,11 @@ pub enum MomoaError {
     #[error("Unexpected end of input found.")]
     UnexpectedEndOfInput {
         loc: Location,
+    }
+}
+
+impl fmt::Debug for MomoaError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str(&self.to_string())
     }
 }
