+++ src/lib.rs
@@ -1,3 +1,7 @@
-pub mod tokens;
-pub mod errors;
-pub mod location;
+mod tokens;
+mod errors;
+mod location;
+
+pub use tokens::{tokenize, TokenKind};
+pub use location::{Location, LocationRange};
+pub use errors::MomoaError;
